refactor: organize custom scripts to its own file
